refactor: rename interface > interfaces
Trying potential solution just in case the singular form could
be interfering the mock generation. Renaming the package to
the plural form.
