+++ internal/api/http/docs.go
@@ -1,0 +1,1 @@
+package http
