+++ internal/handler/event/docs.go
@@ -1,0 +1,1 @@
+package event
