+++ internal/handler/http/impl/application.go
@@ -5,7 +5,7 @@ import (
 	common_err "github.com/avisiedo/go-microservice-1/internal/errors/common"
 	handler_http "github.com/avisiedo/go-microservice-1/internal/handler/http"
 	metrics "github.com/avisiedo/go-microservice-1/internal/infrastructure/metrics"
-	presenter_interface "github.com/avisiedo/go-microservice-1/internal/interface/presenter/sync/echo"
+	presenter_interface "github.com/avisiedo/go-microservice-1/internal/interfaces/presenter/sync/echo"
 	"github.com/avisiedo/go-microservice-1/internal/usecase/interactor"
 	presenter "github.com/avisiedo/go-microservice-1/internal/usecase/presenter/sync/echo"
 	repository "github.com/avisiedo/go-microservice-1/internal/usecase/repository/db"
