+++ internal/infrastructure/docs.go
@@ -1,0 +1,1 @@
+package infrastructure
