+++ internal/interfaces/docs.go
@@ -1,0 +1,1 @@
+package interfaces
