+++ internal/interfaces/interactor/private_interactor.go
