+++ internal/interfaces/presenter/sync/docs.go
@@ -1,0 +1,1 @@
+package sync
