+++ internal/interfaces/presenter/sync/echo/openapi_presenter.go
