+++ internal/interfaces/presenter/sync/echo/todo_presenter_output.go
