+++ internal/interfaces/repository/docs.go
@@ -1,0 +1,1 @@
+package repository
