+++ internal/test/mock/interfaces/interactor/private.go
