+++ internal/usecase/interactor/healthcheck_interactor.go
@@ -2,7 +2,7 @@ package interactor
 
 import (
 	"github.com/avisiedo/go-microservice-1/internal/config"
-	"github.com/avisiedo/go-microservice-1/internal/interface/interactor"
+	"github.com/avisiedo/go-microservice-1/internal/interfaces/interactor"
 )
 
 type healtchcheckInteractor struct {
