+++ internal/usecase/interactor/private_interactor.go
@@ -1,6 +1,6 @@
 package interactor
 
-import "github.com/avisiedo/go-microservice-1/internal/interface/interactor"
+import "github.com/avisiedo/go-microservice-1/internal/interfaces/interactor"
 
 type private struct{}
