+++ internal/usecase/presenter/sync/echo/instrumentation_presenter.go
@@ -3,7 +3,7 @@ package echo
 import (
 	common_err "github.com/avisiedo/go-microservice-1/internal/errors/common"
 	infra_metrics "github.com/avisiedo/go-microservice-1/internal/infrastructure/metrics"
-	presenter "github.com/avisiedo/go-microservice-1/internal/interface/presenter/sync/echo"
+	presenter "github.com/avisiedo/go-microservice-1/internal/interfaces/presenter/sync/echo"
 	"github.com/labstack/echo/v4"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
 )
