+++ internal/usecase/presenter/sync/echo/private_presenter_test.go
@@ -4,7 +4,7 @@ import (
 	"testing"
 
 	common_err "github.com/avisiedo/go-microservice-1/internal/errors/common"
-	"github.com/avisiedo/go-microservice-1/internal/test/mock/interface/interactor"
+	"github.com/avisiedo/go-microservice-1/internal/test/mock/interfaces/interactor"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
