+++ internal/usecase/presenter/sync/echo/todo_presenter_input.go
@@ -6,7 +6,7 @@ import (
 	"github.com/avisiedo/go-microservice-1/internal/api/http/public"
 	"github.com/avisiedo/go-microservice-1/internal/domain/model"
 	common_err "github.com/avisiedo/go-microservice-1/internal/errors/common"
-	. "github.com/avisiedo/go-microservice-1/internal/interface/presenter/sync/echo"
+	. "github.com/avisiedo/go-microservice-1/internal/interfaces/presenter/sync/echo"
 	"github.com/labstack/echo/v4"
 )
