+++ internal/usecase/presenter/sync/echo/todo_presenter_test.go
@@ -12,13 +12,13 @@ import (
 	"github.com/avisiedo/go-microservice-1/internal/config"
 	"github.com/avisiedo/go-microservice-1/internal/domain/model"
 	common_err "github.com/avisiedo/go-microservice-1/internal/errors/common"
-	presenter "github.com/avisiedo/go-microservice-1/internal/interface/presenter/sync/echo"
+	presenter "github.com/avisiedo/go-microservice-1/internal/interfaces/presenter/sync/echo"
 	"github.com/avisiedo/go-microservice-1/internal/test"
 	http_builder "github.com/avisiedo/go-microservice-1/internal/test/builder/api/http"
 	model_builder "github.com/avisiedo/go-microservice-1/internal/test/builder/model"
 	echo_helper "github.com/avisiedo/go-microservice-1/internal/test/helper/http/echo"
-	interactor_mock "github.com/avisiedo/go-microservice-1/internal/test/mock/interface/interactor"
-	presenter_mock "github.com/avisiedo/go-microservice-1/internal/test/mock/interface/presenter/echo"
+	interactor_mock "github.com/avisiedo/go-microservice-1/internal/test/mock/interfaces/interactor"
+	presenter_mock "github.com/avisiedo/go-microservice-1/internal/test/mock/interfaces/presenter/echo"
 	"github.com/google/uuid"
 	"github.com/labstack/echo/v4"
 	openapi_types "github.com/oapi-codegen/runtime/types"
